fix: fresh-clone build, CI on push, JS-interop fractional mtimes
Verification pass against the pushed remake branch surfaced three real
issues, all confirmed by reproduction and fixed:

1. Fresh clones did not compile: frontend/dist was fully gitignored, so
   `go:embed all:frontend/dist` had nothing to match. A committed
   .gitkeep placeholder (with a gitignore exception) keeps clean
   checkouts and CI building; `npm run build` replaces it.

2. No tests ran on branch pushes: release.yml only triggers on tags.
   Tests now live in ci.yml, running vet + race-enabled tests on every
   push/PR across ubuntu and windows.

3. Cross-version sync with the original JS daemon failed on manifest
   decode: Node sends fs.statSync().mtimeMs as FRACTIONAL floats
   (e.g. 1783279365872.0251) which cannot unmarshal into int64. New
   delta.Milli type accepts any JSON number and marshals back as a
   whole integer, with a regression test using the exact JS payload.

Interop was then verified end-to-end on this machine with the real
Node daemon and the Go daemon running side by side: pairing handshake
with approve-confirm both ways, Go->JS file sync (JS pulled blocks
from the Go daemon byte-perfect), and JS->Go (JS watcher detected a
change, triggered the Go daemon, Go pulled and patched). The public
relay at opensave-relay.onrender.com was also health-checked live.

+++ internal/p2p/syncengine/decision.go
@@ -32,8 +32,8 @@ func DetectConflict(local, remote delta.Manifest, lastSyncTimeMs int64) bool {
 	if lastSyncTimeMs == 0 {
 		return len(local.Files) > 0 && len(remote.Files) > 0
 	}
-	localModified := local.LatestMtime > lastSyncTimeMs+clockSkewToleranceMs
-	remoteModified := remote.LatestMtime > lastSyncTimeMs+clockSkewToleranceMs
+	localModified := int64(local.LatestMtime) > lastSyncTimeMs+clockSkewToleranceMs
+	remoteModified := int64(remote.LatestMtime) > lastSyncTimeMs+clockSkewToleranceMs
 	return localModified && remoteModified
 }
